refactor: extract text utils
