+++ src/handlers.rs
@@ -7,6 +7,7 @@ use teloxide::{
 };
 
 use crate::db::*;
+use crate::text_utils::{capitalize_first, parse_item_line};
 
 pub async fn help(bot: Bot, msg: Message) -> Result<()> {
     bot.send_message(
@@ -89,40 +90,6 @@ pub fn format_plain_list(items: &[Item]) -> String {
         text.push_str(&format!("• {}\n", item.text));
     }
     text
-}
-
-/// Clean a single text line from a user message.
-///
-/// Returns `None` if the line should be ignored (for example it is the
-/// archived list separator or becomes empty after trimming). Otherwise returns
-/// the cleaned line without leading status emojis or whitespace.
-pub fn parse_item_line(line: &str) -> Option<String> {
-    tracing::trace!(?line, "Parsing item line");
-    if line.trim() == "--- Archived List ---" {
-        tracing::trace!("Ignoring archived list separator");
-        return None;
-    }
-
-    let cleaned = line
-        .trim_start_matches(['☑', '✅', '⬜', '🛒', '\u{fe0f}'])
-        .trim();
-
-    if cleaned.is_empty() {
-        tracing::trace!("Line empty after cleaning");
-        None
-    } else {
-        let result = cleaned.to_string();
-        tracing::trace!(?result, "Parsed line");
-        Some(result)
-    }
-}
-
-fn capitalize_first(text: &str) -> String {
-    let mut chars = text.chars();
-    match chars.next() {
-        Some(c) => c.to_uppercase().chain(chars).collect(),
-        None => String::new(),
-    }
 }
 
 use crate::ai::stt::{
