+++ src/lib.rs
@@ -7,10 +7,12 @@ use teloxide::{prelude::*, utils::command::BotCommands};
 pub mod ai;
 mod db;
 mod handlers;
+mod text_utils;
 
 pub use ai::stt::{parse_items, parse_voice_items};
 pub use db::Item;
-pub use handlers::{format_delete_list, format_list, format_plain_list, parse_item_line};
+pub use handlers::{format_delete_list, format_list, format_plain_list};
+pub use text_utils::{capitalize_first, parse_item_line};
 
 use handlers::{
     add_items_from_parsed_text, add_items_from_photo, add_items_from_text, add_items_from_voice,
